# Parse whitespace more precisely