--- a/dhall/tests/Dhall/Test/Parser.hs
+++ b/dhall/tests/Dhall/Test/Parser.hs
@@ -89,7 +89,7 @@ notesInLetInLet = do
                       (Just " ")
                       Nothing
                       (Just " ")
-                      (Note "0 " (NaturalLit 0)))
+                      (Note "0" (NaturalLit 0)))
                     -- This 'Let' isn't wrapped in a 'Note'!
                     (Let
                       (Binding
@@ -98,7 +98,7 @@ notesInLetInLet = do
                         (Just " ")
                         Nothing
                         (Just " ")
-                        (Note "1 " (NaturalLit 1))
+                        (Note "1" (NaturalLit 1))
                       )
                       (Note "let z = 2 in x"
                         (Let
@@ -108,7 +108,7 @@ notesInLetInLet = do
                             (Just " ")
                             Nothing
                             (Just " ")
-                            (Note "2 " (NaturalLit 2))
+                            (Note "2" (NaturalLit 2))
                           )
                           (Note "x"
                             (Var (V "x" 0))))))))
@@ -125,8 +125,8 @@ shouldParse path = do
             [ parseDirectory </> "success/unit/MergeParenAnnotation"
 
             -- https://github.com/dhall-lang/dhall-haskell/issues/1454
-            , parseDirectory </> "success/preferMissingNoSpaces"
-            , parseDirectory </> "success/missingInParentheses"
+            , parseDirectory </> "success/missingFoo"
+            , parseDirectory </> "success/missingSlash"
             ]
 
     let pathString = Text.unpack path
@@ -157,35 +157,11 @@ shouldParse path = do
 shouldNotParse :: Text -> TestTree
 shouldNotParse path = do
     let expectedFailures =
-            [ -- These two unexpected successes are due to not correctly
-              -- requiring non-empty whitespace after the `:` in a type
-              -- annotation
-              parseDirectory </> "failure/unit/ImportEnvWrongEscape.dhall"
-
-              -- Other spacing related unexpected successes:
-            , parseDirectory </> "failure/spacing/AnnotationNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ApplicationNoSpace1.dhall"
-            , parseDirectory </> "failure/spacing/ApplicationNoSpace2.dhall"
-            , parseDirectory </> "failure/spacing/AssertNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ForallNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ImportAltNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ImportHashedNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/LambdaNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ListLitEmptyNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/MergeAnnotationNoSpace3.dhall"
-            , parseDirectory </> "failure/spacing/MergeNoSpace2.dhall"
-            , parseDirectory </> "failure/spacing/NaturalPlusNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/RecordTypeNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/ToMapAnnotNoSpace.dhall"
-            , parseDirectory </> "failure/spacing/UnionTypeNoSpace.dhall"
-
-            , parseDirectory </> "failure/ImportHeadersExteriorHash.dhall"
-
-              -- For parsing performance reasons the implementation
+            [ -- For parsing performance reasons the implementation
               -- treats a missing type annotation on an empty list as
               -- as a type-checking failure instead of a parse failure,
               -- but this might be fixable.
-            , parseDirectory </> "failure/unit/ListLitEmptyMissingAnnotation.dhall"
+              parseDirectory </> "failure/unit/ListLitEmptyMissingAnnotation.dhall"
             , parseDirectory </> "failure/unit/ListLitEmptyAnnotation.dhall"
 
               -- The same performance improvements also broke the