This is preparatory work for #1454.