@sjakobi 2019-10-28
So far, the parser is still quite broken. For example:

```
      ./dhall-lang/tests/parser/success/unit/MergeXYZ:                        FAIL
        Exception: 
        Error: Invalid input
        
        1:9:
          |
        1 | merge x y z
          |         ^^
        unexpected "y "
        expecting operator or whitespace
      ./dhall-lang/tests/parser/success/unit/MergeParenAnnotation:            FAIL
        Exception: 
        Error: Invalid input
        
        1:10:
          |
        1 | (merge x y) : t
          |          ^^
        unexpected "y)"
        expecting operator or whitespace
      ./dhall-lang/tests/parser/success/unit/Merge:                           FAIL
        Exception: 
        Error: Invalid input
        
        1:9:
          |
        1 | merge x y
          |         ^^
        unexpected "y<newline>"
        expecting operator or whitespace
      ./dhall-lang/tests/parser/success/unit/ListLitEmptyPrecedence:          FAIL
        Exception: 
        Error: Invalid input
        
        1:11:
          |
        1 | [] : List T U
          |           ^
        unexpected 'T'
        expecting ->, :, end of input, operator, or whitespace
      ./dhall-lang/tests/parser/success/unit/ListLitNonEmptyAnnotated:        FAIL
        Exception: 
        Error: Invalid input
        
        1:15:
          |
        1 | [x, y] : List T
          |               ^
        unexpected 'T'
        expecting ->, :, end of input, operator, or whitespace
```