@Gabriel439 2019-10-29
@sjakobi: I pushed a change with some fixes that illustrates what went wrong.  Many of the remaining fixes are due to the import parser needing to be fixed to not consume trailing whitespace