@sjakobi 2019-10-31
Thanks a lot @Gabriel439! :)

This is the last parser test failure:

```
      ./dhall-lang/tests/parser/success/largeExpression:                                     FAIL
        Exception: 
        Error: Invalid input
        
        67:3:
           |
        67 |   (   λ ( x
           |   ^
        unexpected '('
        expecting end of input or whitespace
```

There are a bunch of other test failures remaining though.