@sjakobi 2019-10-31
A few test failures in dhall-lsp-server:

```
  Completion
    Dhall.Completion
      suggests user defined types:     FAIL (0.03s)
        expected: "Config"
         but got: "toMap"
      suggests user defined functions: FAIL (0.03s)
        expected: "makeUser"
         but got: "toMap"
      suggests user defined bindings:  FAIL (0.03s)
        expected: "bob"
         but got: "toMap"
      suggests functions from imports: FAIL (0.03s)
        uncaught exception: ErrorCall
        Prelude.head: empty list
  Hovering
    Dhall.Hover
      reports types on hover:          FAIL (0.03s)
        expected: "{ home : Text, name : Text }"
         but got: "Type"
```