@sjakobi 2019-11-01
@EggBaconAndSpam Do you have some advice on how to best update `dhall-lsp-server`? Should I attempt to fix the parsers in [`Dhall.LSP.Backend.Parsing`](https://github.com/dhall-lang/dhall-haskell/blob/master/dhall-lsp-server/src/Dhall/LSP/Backend/Parsing.hs) in analogy to the changes in `Dhall.Parser.Expression` or could I potentially completely get rid of them now?