@sjakobi 2019-11-01
It took me _way_ to long to figure out why my changes to `dhall-lsp-server` wouldn't change the test results from `stack test dhall-lsp-server:tests`:

1. `dhall-lsp-server:tests` is an integration test suite that depends on the _executable_ `dhall-lsp-server`.
2. While `stack` happily reports

   > Installing executable dhall-lsp-server in <path>

    every time I change the library, it will only re-compile the executable if it's one of the targets!

So you have to include the executable in the targets and run

    stack test dhall-lsp-server:tests dhall-lsp-server:dhall-lsp-server

:scream: :scream: :scream: