@sjakobi 2019-11-01
I've modified `dhall-lsp-server`'s just enough now that the tests pass. `dhall-lsp-server` is very likely still broken for other inputs though.