@sjakobi 2019-11-02
While looking at `dhall-lsp-server`'s tests I was surprised by this expectation:

https://github.com/dhall-lang/dhall-haskell/blob/1b46f1885f25c16e369b3f455c3950e4837ac49f/dhall-lsp-server/tests/Main.hs#L40

My understanding is that this should be the type of `mkUser` here:

https://github.com/dhall-lang/dhall-haskell/blob/1b46f1885f25c16e369b3f455c3950e4837ac49f/dhall-lsp-server/tests/fixtures/hovering/Types.dhall#L3-L9

Shouldn't the type be

    Bool -> { home : Text, name : Text }

then?

(CC @mujx)