@Gabriel439 2019-11-02
@sjakobi: It's a bug in `dhall-lsp-server`.  I can reproduce:

<img width="268" alt="Screen Shot 2019-11-02 at 11 44 48 AM" src="https://user-images.githubusercontent.com/1313787/68075546-46edec80-fd66-11e9-8de1-26ed679ffd6b.png">