@sjakobi 2019-11-03
@Gabriel439 I have made a new issue for this: https://github.com/dhall-lang/dhall-haskell/issues/1510

Would you mind giving this a final review, so we can get this merged?