@sjakobi 2019-11-03
I remembered that we should look at the performance impact:

I've benchmarked `dhall resolve --immediate-dependencies` for `cpkg`'s `pkg-set.dhall`:

#### `master`

```
$ bench "dhall resolve --immediate-dependencies --file pkgs/pkg-set.dhall"
benchmarking dhall resolve --immediate-dependencies --file pkgs/pkg-set.dhall
time                 154.7 ms   (151.9 ms .. 156.1 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 155.9 ms   (154.9 ms .. 156.7 ms)
std dev              1.217 ms   (701.7 μs .. 1.891 ms)
variance introduced by outliers: 12% (moderately inflated)
```

#### This branch

```
$ bench "dhall resolve --immediate-dependencies --file pkgs/pkg-set.dhall"
benchmarking dhall resolve --immediate-dependencies --file pkgs/pkg-set.dhall
time                 232.9 ms   (226.7 ms .. 246.0 ms)
                     0.999 R²   (0.993 R² .. 1.000 R²)
mean                 228.1 ms   (225.2 ms .. 233.7 ms)
std dev              4.848 ms   (1.233 ms .. 6.609 ms)
variance introduced by outliers: 14% (moderately inflated)
```

This is an increase of 50.5%!

In the micro-benchmarks I've noticed particular increases for the comment parsing:

https://github.com/dhall-lang/dhall-haskell/blob/cc1814be9858f62e664b0d2114e711fd4431bd2c/dhall/benchmark/parser/Main.hs#L95-L96

#### master

```
benchmarked Line comment
time                 11.86 ms   (11.69 ms .. 11.98 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 11.84 ms   (11.79 ms .. 11.89 ms)
std dev              129.4 μs   (107.2 μs .. 164.1 μs)

benchmarked Block comment
time                 13.20 ms   (13.00 ms .. 13.41 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 13.59 ms   (13.41 ms .. 13.94 ms)
std dev              600.0 μs   (142.2 μs .. 953.7 μs)
variance introduced by outliers: 15% (moderately inflated)
```

#### This branch

```
benchmarked Line comment
time                 228.6 ms   (225.6 ms .. 231.1 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 230.3 ms   (228.9 ms .. 233.8 ms)
std dev              3.631 ms   (1.668 ms .. 5.820 ms)

benchmarking Block comment ... took 14.31 s, total 56 iterations
benchmarked Block comment
time                 260.9 ms   (234.1 ms .. 298.2 ms)
                     0.976 R²   (0.940 R² .. 0.997 R²)
mean                 255.3 ms   (243.2 ms .. 273.0 ms)
std dev              25.54 ms   (17.55 ms .. 35.68 ms)
variance introduced by outliers: 29% (moderately inflated)
```

This got about 20x slower!